cleanup: reorganise gl code
